requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
sentence-transformers==3.2.1
jupyterlab==4.4.5
notebook==7.4.4
ipykernel==6.27.0
tqdm==4.67.1
python-dotenv==1.0.1
google-generativeai==0.8.5
